feat(tui): extract ChatScreen with event-driven status (GH13 Phase 3)
- Create ChatScreen with onEvent callback for real-time status indicators
  (Thinking/Running code/Code finished/error)
- Persist messages to store and load on mount
- Refactor App.tsx to thin shell delegating to ChatScreen
- Remove ReviewPage.tsx (absorbed into upcoming ToolsScreen)
- Pass sessionId through to agent.chat for session-title generation
